+++ src/test/kotlin/com/github/gabortim/phonenumber/tool/NumberFormatterTest.kt
@@ -7,37 +7,28 @@ import com.github.gabortim.phonenumber.tool.NumberFormatter.possiblyValidExtensi
 import com.google.i18n.phonenumbers.NumberParseException
 import com.google.i18n.phonenumbers.PhoneNumberUtil
 import com.google.i18n.phonenumbers.Phonenumber.PhoneNumber
-import org.testng.Assert.assertEquals
-import org.testng.Assert.assertFalse
-import org.testng.Assert.assertTrue
-import org.testng.Assert.fail
-import org.testng.annotations.AfterMethod
-import org.testng.annotations.BeforeMethod
-import org.testng.annotations.Test
+import org.junit.jupiter.api.Assertions.*
+import org.junit.jupiter.api.Disabled
+import org.junit.jupiter.api.Test
+import org.openstreetmap.josm.testutils.annotations.BasicPreferences
 
+
+@BasicPreferences
 class NumberFormatterTest {
     companion object {
-        private val phoneNumberUtil: PhoneNumberUtil = PhoneNumberUtil.getInstance()
-    }
-
-    @BeforeMethod
-    fun setUp() {
-    }
-
-    @AfterMethod
-    fun tearDown() {
+        private val util: PhoneNumberUtil = PhoneNumberUtil.getInstance()
     }
 
     @Test
     fun testFormatNumber_AT() {
         val region = "AT"
         try {
-            var number = phoneNumberUtil.parseAndKeepRawInput("+43 50 5333-6630", region)
-            assertEquals(format(number).first, "+43 50 5333 ext. 6630")
-            number = phoneNumberUtil.parseAndKeepRawInput("+43 5675 6444-26", region)
-            assertEquals(format(number).first, "+43 5675 6444 ext. 26")
-            number = phoneNumberUtil.parseAndKeepRawInput("+43 5558 8213-410", region)
-            assertEquals(format(number).first, "+43 5558 8213 ext. 410")
+            var number = util.parseAndKeepRawInput("+43 50 5333-6630", region)
+            assertEquals("+43 50 5333 ext. 6630", format(number).first)
+            number = util.parseAndKeepRawInput("+43 5675 6444-26", region)
+            assertEquals("+43 5675 6444 ext. 26", format(number).first)
+            number = util.parseAndKeepRawInput("+43 5558 8213-410", region)
+            assertEquals("+43 5558 8213 ext. 410", format(number).first)
         } catch (e: NumberParseException) {
             e.printStackTrace()
             fail()
@@ -48,8 +39,8 @@ class NumberFormatterTest {
     fun testFormatNumber_DE() {
         val region = "DE"
         try {
-            val number = phoneNumberUtil.parseAndKeepRawInput("+49 771 898649-9", region)
-            assertEquals(format(number).first, "+49 771 898649 ext. 9")
+            val number = util.parseAndKeepRawInput("+49 771 898649-9", region)
+            assertEquals("+49 771 898649 ext. 9", format(number).first)
         } catch (e: NumberParseException) {
             e.printStackTrace()
             fail()
@@ -60,35 +51,35 @@ class NumberFormatterTest {
     fun testFormatNumber_HU() {
         val region = "HU"
         try {
-            var number = phoneNumberUtil.parseAndKeepRawInput("+36 62 800800", region)
-            assertEquals(format(number!!).first, "+36 62 800 800")
+            var number = util.parseAndKeepRawInput("+36 62 800800", region)
+            assertEquals("+36 62 800 800", format(number!!).first)
 
-            number = phoneNumberUtil.parseAndKeepRawInput("+3662488446", region)
-            assertEquals(format(number).first, "+36 62 488 446")
+            number = util.parseAndKeepRawInput("+3662488446", region)
+            assertEquals("+36 62 488 446", format(number).first)
 
-            number = phoneNumberUtil.parseAndKeepRawInput("(62)426483", region)
-            assertEquals(format(number).first, "+36 62 426 483")
+            number = util.parseAndKeepRawInput("(62)426483", region)
+            assertEquals("+36 62 426 483", format(number).first)
 
-            number = phoneNumberUtil.parseAndKeepRawInput("+36209209000", region)
-            assertEquals(format(number).first, "+36 20 920 9000")
+            number = util.parseAndKeepRawInput("+36209209000", region)
+            assertEquals("+36 20 920 9000", format(number).first)
 
-            number = phoneNumberUtil.parseAndKeepRawInput("+36209209000", region)
-            assertEquals(format(number).first, "+36 20 920 9000")
+            number = util.parseAndKeepRawInput("+36209209000", region)
+            assertEquals("+36 20 920 9000", format(number).first)
 
-            number = phoneNumberUtil.parseAndKeepRawInput("+36 62/424-805 / 612", region)
-            assertEquals(format(number).first, "+36 62 424 805 ext. 612")
+            number = util.parseAndKeepRawInput("+36 62/424-805 / 612", region)
+            assertEquals("+36 62 424 805 ext. 612", format(number).first)
 
-            number = phoneNumberUtil.parseAndKeepRawInput("+36 62/424-805 # 612", region)
-            assertEquals(format(number).first, "+36 62 424 805 ext. 612")
+            number = util.parseAndKeepRawInput("+36 62/424-805 # 612", region)
+            assertEquals("+36 62 424 805 ext. 612", format(number).first)
 
-            number = phoneNumberUtil.parseAndKeepRawInput("+36 62/424-805 - 612", region)
-            assertEquals(format(number).first, "+36 62 424 805 ext. 612")
+            number = util.parseAndKeepRawInput("+36 62/424-805 - 612", region)
+            assertEquals("+36 62 424 805 ext. 612", format(number).first)
 
-            number = phoneNumberUtil.parseAndKeepRawInput("+36 1 3408132", region)
-            assertEquals(format(number).first, "+36 1 340 8132")
+            number = util.parseAndKeepRawInput("+36 1 3408132", region)
+            assertEquals("+36 1 340 8132", format(number).first)
 
-            number = phoneNumberUtil.parseAndKeepRawInput("1272", region)
-            assertEquals(format(number).first, "1272")
+            number = util.parseAndKeepRawInput("1272", region)
+            assertEquals("1272", format(number).first)
 
         } catch (e: NumberParseException) {
             e.printStackTrace()
@@ -100,11 +91,11 @@ class NumberFormatterTest {
     fun testFormatNumber_US() {
         val region = "US"
         try {
-            var number = phoneNumberUtil.parseAndKeepRawInput("931-647-1567", region)
-            assertEquals(format(number).first, "+1 931-647-1567")
+            var number = util.parseAndKeepRawInput("931-647-1567", region)
+            assertEquals("+1 931-647-1567", format(number).first)
 
-            number = phoneNumberUtil.parseAndKeepRawInput("1-615-371-9200", region)
-            assertEquals(format(number).first, "+1 615-371-9200")
+            number = util.parseAndKeepRawInput("1-615-371-9200", region)
+            assertEquals("+1 615-371-9200", format(number).first)
         } catch (e: NumberParseException) {
             e.printStackTrace()
             fail()
@@ -112,23 +103,25 @@ class NumberFormatterTest {
     }
 
     @Test
-    @Throws(NumberParseException::class)
     fun testNoFormatNumber() {
         val region = "HU"
         val s = "(Polgármester: Lehoczki Szabolcs, 2019- )"
-        val number = phoneNumberUtil.parseAndKeepRawInput(s, region)
-        format(number)
+        val number = util.parseAndKeepRawInput(s, region)
+        assertEquals("", format(number).first)
     }
 
-    @Test(enabled = false)
+    @Test
+    @Disabled
     fun testGetGroupingCharsCount() {
     }
 
-    @Test(enabled = false)
+    @Test
+    @Disabled
     fun testIsValid() {
     }
 
-    @Test(enabled = false)
+    @Test
+    @Disabled
     fun testIsValidShort() {
     }
 
@@ -136,9 +129,9 @@ class NumberFormatterTest {
     fun testIsPremiumRateNumber() {
         val region = "HU"
         try {
-            var number = phoneNumberUtil.parseAndKeepRawInput("+36 90 317 282", region)
+            var number = util.parseAndKeepRawInput("+36 90 317 282", region)
             assertTrue(isPremiumRate(number!!))
-            number = phoneNumberUtil.parseAndKeepRawInput("+36 91 500 316", region)
+            number = util.parseAndKeepRawInput("+36 91 500 316", region)
             assertTrue(isPremiumRate(number))
         } catch (e: NumberParseException) {
             e.printStackTrace()
@@ -151,7 +144,7 @@ class NumberFormatterTest {
         val region = "HU"
         val number: PhoneNumber
         try {
-            number = phoneNumberUtil.parseAndKeepRawInput("112", region)
+            number = util.parseAndKeepRawInput("112", region)
             assertTrue(isEmergency(number))
         } catch (e: NumberParseException) {
             e.printStackTrace()