+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2023-10-XX
+### Added
+- More tests, including some regression tests
+### Changed
+- Update Gradle to 8.4
+- Switch the test framework from TestNG to JUnit because of upstream dependency
+- Fixed a typo in a warning message
+- Fixed duplicate warning for the same autofixable issue (bad separator and beatutifiable)
+
 ## [1.1.0] - 2023-10-05
 ### Added
 - New validator message in case no (auto-fixable) formatting issue found